pyproject.toml
[project]
name = "python-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.3",
    "pyexecjs>=1.5.1",
    "requests>=2.32.4",
]
